Moves the surviving Home functionality into the Commit Graph
- Relocates the account and integrations chips beside their only
remaining
  consumer, the Graph account rollup
- Explains Git-blocked unsafe repositories in the Graph's empty state
with a
  Manage in Source Control action, replacing the generic no-repository
message

(#5778, #5808)
