+++ src/webviews/plus/graph/graphWebview.ts
@@ -5376,6 +5376,7 @@ export class GraphWebviewProvider implements WebviewProvider<State, State, Graph
 				...this.host.baseWebviewState,
 				allowed: true,
 				trusted: true,
+				hasUnsafeRepositories: this.container.git.hasUnsafeRepositories(),
 				repositories: [],
 				isWeb: isWeb,
 				subscription: subscription,
@@ -5390,6 +5391,7 @@ export class GraphWebviewProvider implements WebviewProvider<State, State, Graph
 					...this.host.baseWebviewState,
 					allowed: true,
 					trusted: true,
+					hasUnsafeRepositories: this.container.git.hasUnsafeRepositories(),
 					repositories: [],
 					isWeb: isWeb,
 					subscription: subscription,
